+++ src/main/java/nextstep/web/ReservationController.java
@@ -1,36 +1,37 @@
 package nextstep.web;
 
 import nextstep.model.Reservation;
-import nextstep.web.dto.ReservationRequest;
-import nextstep.web.dto.ReservationResponse;
+import nextstep.service.ReservationService;
+import nextstep.dto.ReservationRequest;
+import nextstep.dto.ReservationResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 
 @RestController
-public class RoomEscapeController {
-    private final RoomEscapeService roomEscapeService;
+public class ReservationController {
+    private final ReservationService reservationService;
 
-    public RoomEscapeController(RoomEscapeService roomEscapeService) {
-        this.roomEscapeService = roomEscapeService;
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @PostMapping("/reservations")
     public ResponseEntity<Void> createReservation(@RequestBody ReservationRequest request) {
-        Reservation reservation = roomEscapeService.createReservation(request);
+        Reservation reservation = reservationService.createReservation(request);
         return ResponseEntity.created(URI.create("/reservations/"+ reservation.getId())).build();
     }
 
     @GetMapping("/reservations/{id}")
     public ResponseEntity<ReservationResponse> getReservation(@PathVariable Long id) {
-        Reservation reservation = roomEscapeService.getReservation(id);
+        Reservation reservation = reservationService.getReservation(id);
         return ResponseEntity.ok(new ReservationResponse(reservation));
     }
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        roomEscapeService.deleteReservation(id);
+        reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
     }
 }
